added roles system

+++ api/handlers.py
@@ -1,15 +1,19 @@
 from uuid import UUID
+from logging import getLogger
 
 from fastapi import APIRouter, Depends, HTTPException
+from sqlalchemy.exc import IntegrityError
 from sqlalchemy.ext.asyncio import AsyncSession
 
 from api.actions.auth import get_current_user_from_token
-from api.actions.user import create_new_user, delete_existing_user, get_existing_user_by_id, update_user
+from api.actions.user import create_new_user, delete_existing_user, get_existing_user_by_id, update_user, \
+    check_user_permissions
 from api.models import UserCreate, ShowUser, DeleteUserResponse, UpdatedUserResponse, UpdateUserRequest
 from db.models import User
 from db.session import get_db
 
 user_router = APIRouter()
+logger = getLogger(__name__)
 
 
 @user_router.post("/", response_model=ShowUser)
@@ -20,9 +24,21 @@ async def create_user(body: UserCreate, db: AsyncSession = Depends(get_db)) -> S
 @user_router.delete("/", response_model=DeleteUserResponse)
 async def delete_user(user_id: UUID, db: AsyncSession = Depends(get_db),
                       current_user: User = Depends(get_current_user_from_token), ) -> DeleteUserResponse:
+    user_for_deletion = await get_existing_user_by_id(user_id, db)
+    if user_for_deletion is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
+    if not check_user_permissions(
+            target_user=user_for_deletion,
+            current_user=current_user,
+    ):
+        raise HTTPException(status_code=403, detail="Forbidden.")
     deleted_user_id = await delete_existing_user(user_id, db)
     if deleted_user_id is None:
-        raise HTTPException(status_code=404, detail=f"User with id {user_id} not found.")
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
     return DeleteUserResponse(deleted_user_id=deleted_user_id)
 
 
@@ -36,14 +52,29 @@ async def get_user_by_id(user_id: UUID, db: AsyncSession = Depends(get_db),
 
 
 @user_router.patch("/", response_model=UpdatedUserResponse)
-async def update_user_by_id(
-        user_id: UUID, body: UpdateUserRequest, db: AsyncSession = Depends(get_db),
-        current_user: User = Depends(get_current_user_from_token), ) -> UpdatedUserResponse:
+async def update_user_by_id(user_id: UUID, body: UpdateUserRequest, db: AsyncSession = Depends(get_db),
+                            current_user: User = Depends(get_current_user_from_token), ) -> UpdatedUserResponse:
     updated_user_params = body.dict(exclude_none=True)
     if updated_user_params == {}:
-        raise HTTPException(status_code=422, detail="At least one parameter for user update info should be provided")
-    user = await get_existing_user_by_id(user_id, db)
-    if user is None:
-        raise HTTPException(status_code=404, detail=f"User with id {user_id} not found.")
-    updated_user_id = await update_user(updated_user_params=updated_user_params, session=db, user_id=user_id)
+        raise HTTPException(
+            status_code=422,
+            detail="At least one parameter for user update info should be provided",
+        )
+    user_for_update = await get_existing_user_by_id(user_id, db)
+    if user_for_update is None:
+        raise HTTPException(
+            status_code=404, detail=f"User with id {user_id} not found."
+        )
+    if user_id != current_user.user_id:
+        if check_user_permissions(
+                target_user=user_for_update, current_user=current_user
+        ):
+            raise HTTPException(status_code=403, detail="Forbidden.")
+    try:
+        updated_user_id = await update_user(
+            updated_user_params=updated_user_params, session=db, user_id=user_id
+        )
+    except IntegrityError as err:
+        logger.error(err)
+        raise HTTPException(status_code=503, detail=f"Database error: {err}")
     return UpdatedUserResponse(updated_user_id=updated_user_id)
